==> Achievement 4/03 Scripts/02 Importing libraries and working with Python Data Types-2.ipynb <==
# Content List

01 Importing Libraries
02 Working with Python Data Types

# 01. Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os


# 02. Working with Python Data types

Working with numerical variables

x = 50

x + 50

x = 99

x + 1

x = 200

x - 100

Working with Floats

x = 45.19

y = 5.25

x / y

Working with strings

a = 'H'

b = 'e'

c = 'l'

d = 'l'

e = 'o'

a + b + c + d + e

a = 'Good Morning,'

b = 'How are you?'

a + b
