==> Achievement 4/03 Scripts/03 IC Data Import and Descriptive Analysis.ipynb <==
# Content List

01 Importing Libraries
02 Importing Data

# 01 Importing Libraries 

# Import libraries
import pandas as pd
import numpy as np
import os


# 02 Importing Data  

path = r'/Users/a/Desktop/4'

path

vars_list = ('order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order')

vars_list

# importing orders.csv 
df = pd.read_csv(os.path.join(path, 'orders.csv'), usecols = vars_list)

df.head()

df = pd.read_csv(os.path.join(path, 'products.csv'), index_col = False)

df.head(20)

df.tail(35)

df.shape

df.columns

Question 01: What’s the maximum value of the “aisle_id” column?
Answer: The Maximum value of the aisle_id column is 134.000000

df.describe()

Question 2:What’s the data type of the “department_id” column?
Answer 2: the data type is int64

df.dtypes
