==> Achievement 4/03 Scripts/04 Data Wrangling & Subsetting.ipynb <==
# Content List

01 Importing Libraries
02 Importing Dataframes
03 Data Wrangling Procedures
04 Exercises 2-14

# 01 IMPORTING LIBRARIES

import pandas as pd
import numpy as np
import os

# 02 IMPORTING DATAFRAMES

path = r'/Users/a/Desktop/4'



path

df_ords = pd.read_csv(os.path.join(path, 'orders.csv'), index_col = False)


df_prods = pd.read_csv(os.path.join(path, 'products.csv'), index_col = False)

df_dep = pd.read_csv(os.path.join(path, 'departments.csv'), index_col = False)

EXERCISE WALK TROUGH

# 03 DATA WRANGLING PROCEDURES

# Drop the eval_set column from orders.csv & redefining the df_ords data frame 

df_ords = df_ords.drop(columns = ['eval_set'])

# Count NaN values in the days_since_prior_order column

df_ords['days_since_prior_order'].value_counts(dropna = False)

# Rename the order_dow column into orders_day_of_week

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)



df_ords.head()

# Changing data type

df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

# transposing departments.csv

df_dep.head()

df_dep.T

#creating a new dataframe from the transposed version
df_dep_t = df_dep.T



df_dep_t

df_dep_t.reset_index()

#Creating a new header
new_header = df_dep_t.iloc[0]

new_header

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

df_dep_t_new.columns = new_header



df_dep_t_new

#Creating a data dictionary

data_dict = df_dep_t_new.to_dict('index')



data_dict

# Using data dictionary

df_prods.head()

# Findin out in which department chocholate cookies belong to
print(data_dict.get('19')) 

# Subsetting

# Create a data frame that only contains data from the snacks department

df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks

# Using loc function

df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]


df_snacks_2

# Checking  multiple departments at once with a list
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])] 

# 04 EXERCISES 2-14

# 2. Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.


df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

# 3. Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the data frame.

df_ords.rename(columns = {'order_hour_of_day' : 'time_of_orders'}, inplace = True)



df_ords.head()

# 4.Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.
--> Busiest hour is 10am

df_ords['time_orders'].value_counts(dropna = False) 

# 5. Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.


print(data_dict.get('19')) 


# 6. The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.


df_breakfast =  df_prods[df_prods['department_id']==14]


df_breakfast.head()

# 7. They’d also like to see details about customers who might be throwing dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.


df_dinner_party = df_prods.loc[df_prods['department_id'].isin([5, 20, 7, 12])] 

df_dinner_party.head()

# 8. Count of last Dataframe

df_dinner_party.shape

# 9. Extracting infromation about user_id = 1

df_user_id1 = df_ords[df_ords['user_id']=='1']

df_user_id1.describe()

# 12. Exporting df_ords

path = r'/Users/a/Career Foundry/Achievement 4'

path

df_ords.to_csv(os.path.join(path, '02 Data','02 Prepared Data','orders_wrangled.csv'))

# 13.Exporting df_dep

df_ords.to_csv(os.path.join(path, '02 Data','02 Prepared Data','departments_wrangled.csv'))
