==> Achievement 4/03 Scripts/05 Data Consistency Checks.ipynb <==
# Content List


01 Importing Libraries
02 Importing Dataframes
03 Mixed Data Type
04 Finding and Addressing Missing Values
05 Finding and Addressing Duplicates
06 Exercises 2 - 9 

# 01 Importing Libraries

import pandas as pd
import numpy as np
import os

# 02 Importing Dataframes

path = r'/Users/a/Desktop/4'



path

df_prods = pd.read_csv(os.path.join(path, 'products.csv'), index_col = False)

path = r'/Users/a/Career Foundry/Achievement 4/02 Data/02 Prepared Data'

path

df_ords = pd.read_csv(os.path.join(path, 'orders_wrangled.csv'), index_col = False)

df_ords.head()

# 03 Mixed Type Data

# Create a dataframe
df_test = pd.DataFrame() 

# Create a mixed type column
df_test['mix'] = ['a','b', 1, True]

df_test.head()

# Check for mixed data types

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

df_test['mix'] = df_test['mix'].astype('str')


# 04 Finding and Addressing Missing Values

# Finding missing values
df_prods.isnull().sum()

df_nan = df_prods[df_prods['product_name'].isnull() == True] 


df_nan

df_prods.shape

# Addressing missing values
# creating a new dataframe
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]


df_prods_clean.shape

# Creating a boolean column 

df_prods['ordered_today'] = np.where(df_prods['days_prior_order'].isnull(), True, False)

# 05 Finding and Addressing Duplicates

# Finding Duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()] 


df_dups

df_prods_clean.shape

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# 06 Exercises 2 - 9 

df_prods.describe()

The maximum price under the prices column is unusually high for a grocery product

# 3. Check for mixed-type data in your df_ords dataframe.


for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)
    


df_ords.head()

4. If you find mixed-type data, fix it. The column in question should contain observations of a single data type.

There are no mixed data types columns

# 5. Run a check for missing values in your df_ords dataframe.
df_ords.isnull().sum()


The column days since prior order has 206209 missing values

df_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True] 


df_nan

Meaning: the customer has placed an order at one day so there are a lot of missing entries for the days before the order was places

# 6. Address the missing values using an appropriate method.


We can´t impute the data because it is string data. I don´t find this specific variable relevant for the analysis, I would simply entry the data of when the order has been placed. In that case, I would solve the issue of missing data. Another way is to create a boolean flaf for orders that have been placed today

df_ords['ordered_today']= np.where(df_ords['days_since_prior_order'].isnull(), True, False)

df_ords.shape

# 7. Run a check for duplicate values in your df_ords data.


df_ords_clean = df_ords[df_ords['days_since_prior_order'].isnull() == False]




df_ords_dups = df_ords_clean[df_ords_clean.duplicated()]

df_ords_dups

df_ords_dups.shape

There are no duplicates

# 9. Export your final, cleaned df_prods and df_ords data as “.csv” files in your “Prepared Data” folder and give them appropriate, succinct names.


path = r'/Users/a/Career Foundry/Achievement 4'

path

df_ords_clean.to_csv(os.path.join(path, '02 Data','02 Prepared Data','orders_cleaned.csv'))

df_prods_clean.to_csv(os.path.join(path, '02 Data','02 Prepared Data','products_cleaned.csv'))
