==> Achievement 4/03 Scripts/06 Combining Data Exercise Walk Through.ipynb <==
# Content List



01 Importing Libraries
02 Importing Dataframes
03 Merging Dataframes
04 Exporting Dataframes

# 01 Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os


# 02 Importing Dataframes

path = r'/Users/a/Desktop/4'

path

df_ords2 = pd.read_csv(os.path.join(path, 'order_prior.csv'), index_col = False)

path = r'/Users/a/Career Foundry/Achievement 4/02 Data/02 Prepared Data'

path

df_ords = pd.read_csv(os.path.join(path,'orders_cleaned.csv' ), index_col = False)

df_ords

df_ords2.head()

df_ords.head()

df_ords2.shape

df_ords.shape

# 03 Merging Dataframes

df_merged_large = df_ords.merge(df_ords2, on = 'order_id', how = 'outer', indicator = True)

df_merged_large['_merge'].value_counts()

# 04 Exporting Dataframe

path = r'/Users/a/Career Foundry/Achievement 4/02 Data/02 Prepared Data'

path

df_merged_large.to_pickle(os.path.join(path,'orders_products_combined.pkl'))



df_merged_large.head()
