==> Achievement 4/03 Scripts/06 Combining Data Exercise.ipynb <==
# Content List



01 Importing Libraries
02 Importing Dataframes
03 Merging 
04 Exporting 

# 01 Importing Libraries

import pandas as pd
import numpy as np
import os

# 02 Importing Dataframe

path = r'/Users/a/Career Foundry/Achievement 4/02 Data/02 Prepared Data'

path 

df_orders_products_combined = pd.read_pickle(os.path.join(path,'orders_products_combined.pkl'))

df_orders_products_combined.head()

# Checking shape of dataframe
df_orders_products_combined.shape

# Importing products dataframe

df_products = pd.read_csv(os.path.join(path,'products_cleaned.csv'), index_col = False)

df_products.head()

df_products.shape

# 03 Merging

# Merging products df with df_orders_products_combined

df_orders_products_merged = df_orders_products_combined.merge(df_products, on = ['product_id'])

df_orders_products_merged.head()

df_orders_products_merged['_merge'].value_counts()

# 04 Exporting

path = r'/Users/a/Career Foundry/Achievement 4/02 Data/02 Prepared Data'

path

df_merged.to_csv(os.path.join(path,'orders_products_merged.csv'))
