==> Achievement 4/03 Scripts/07 Deriving new variable-2.ipynb <==
# Content List



01 Importing Libraries
02 Importing Dataframes
03 If Statement
04 If-Statements with the loc() Function
05 If-Statements with For-Loops
06 Exercise




# 01 Importing Libraries

import pandas as pd
import numpy as np
import os

# 02 Importing Dataframe

path = r'/Users/a/Career Foundry/Achievement 4/02 Data/02 Prepared Data'

path

df_ords = pd.read_csv(os.path.join(path, 'orders_products_merged.csv'), index_col = False)

df = df_ords[:1000000]

# 03 If Statement

# Define a function

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Applying the function

df['price_range'] = df.apply(price_label, axis=1)


df['price_range'].value_counts(dropna = False)

df['prices'].max()

# 04 If-Statements with the loc() Function

df_ords.loc[df_ords['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords.loc[(df_ords['prices'] <= 15) & (df_ords['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 


df_ords.loc[df_ords['prices'] <= 5, 'price_range_loc'] = 'Low-range product'


df_ords['price_range_loc'].value_counts(dropna = False)

# 05 If-Statements with For-Loops


df_ords['orders_day_of_week'].value_counts(dropna = False)

result = []

for value in df_ords["orders_day_of_week"]:
  if value == 0:
    result.append("busiest day")
  elif value == 4:
    result.append("least busy")
  else:
    result.append("regularly busy")

result

df_ords["busiest_day"] = result

df_ords["busiest_day"].value_counts(dropna = False)

# 06 Exercise

# Creating a new column for busiest_days

result = []

for value in df_ords["orders_day_of_week"]:
  if value == 0 & 1:
    result.append("Busiest days")
  elif value == 3 & 4:
    result.append("Slowest days")
  else:
    result.append("Regularly busy")

result

# Creating a new column busiest_days

df_ords["busiest_days"] = result

# Printing frequency of busiest_days

df_ords["busiest_days"].value_counts(dropna = False)

df_ords.head()

# Creating a new column for busiest_period_of_day

# Frequency of time_orders to find out the which hour is the busiest

df_ords['time_orders'].value_counts(dropna = False)

The busiest hour for orders is 10 am

result = []

for value in df_ords["time_orders"]:
  if value == 10:
    result.append("Most orders")
  elif value == 19:
    result.append("Average orders")
  else:
    result.append("Fewest orders")

result

# Creating a new column 

df_ords["busiest_period_of_day"] = result

# Printing frequency of busiest_period_of_day

df_ords["busiest_period_of_day"].value_counts(dropna = False)

# Exporting altered dataframe

path = r'/Users/a/Career Foundry/Achievement 4/02 Data/02 Prepared Data'


path

df_ords.to_pickle(os.path.join(path,'orders_products_altered.pkl'))
